fixed all warnings of -Wnarrowing type
Import internal commit 706d465cc2419b73763d6266353c74da801079af

+++ groups/bsl/bsls/bsls_atomicoperations.t.cpp
@@ -1844,13 +1844,13 @@ int main(int argc, char *argv[]) {
                 int d_amount;      // Amount to add
                 int d_expected;    // Expected value
             } VALUES[] = {
-                //line d_base    d_amount d_expected
-                //---- --------  -------- ----------
-                { L_,   0       , -9    , -9         },
-                { L_,   1       , 0     , 1          },
-                { L_,  -1       , 1     , 0          },
-                { L_, 0xFFFFFFFF, 1     , 0          },
-                { L_,  -2       , -2    , -4         }
+                //line d_base          d_amount d_expected
+                //---- --------        -------- ----------
+                { L_,   0             , -9    , -9         },
+                { L_,   1             , 0     , 1          },
+                { L_,        -1       , 1     , 0          },
+                { L_, (int) 0xFFFFFFFF, 1     , 0          },
+                { L_,        -2       , -2    , -4         }
             };
 
             const int NUM_VALUES = sizeof VALUES / sizeof *VALUES;
@@ -2198,7 +2198,13 @@ int main(int argc, char *argv[]) {
                 { L_,   0   , 11         , 33     , 0          , 0       },
                 { L_,   1   , 19         , 1      , 19         , 1       },
                 { L_,  -1   , 4          , 1      , -1         , -1      },
-                { L_,   2   , 0xFFFFFFFF , 2      , 0xFFFFFFFF , 2       },
+                { L_,
+                  2,
+                  (int) 0xFFFFFFFF,
+                  2,
+                  (int) 0xFFFFFFFF,
+                  2
+                },
                 { L_,  -2   , 16         , 0      , -2         , -2      }
             };
 
@@ -3201,13 +3207,13 @@ int main(int argc, char *argv[]) {
                 bsls::Types::Int64 d_value; // Input value
 
             } VALUES[] = {
-                //line expected             value
-                //---- -------------------  ---------------------
-                { L_,   0                  , 1                    },
-                { L_,   1                  , 2                    },
-                { L_,  -1LL                , 0                    },
-                { L_,   0xFFFFFFFFLL       , 0x100000000LL        },
-                { L_,  0xFFFFFFFFFFFFFFFFLL , 0                   }
+                //line expected                   value
+                //---- -------------------        ---------------------
+                { L_,   0                        , 1                    },
+                { L_,   1                        , 2                    },
+                { L_,  -1LL                      , 0                    },
+                { L_,   0xFFFFFFFFLL             , 0x100000000LL        },
+                { L_,  (int) 0xFFFFFFFFFFFFFFFFLL , 0                   }
             };
 
             const int NUM_VALUES = sizeof VALUES / sizeof *VALUES;
@@ -3610,13 +3616,13 @@ int main(int argc, char *argv[]) {
                 bsls::Types::Int64 d_value; // Input value
 
             } VALUES[] = {
-                //line expected             value
-                //---- -------------------  ---------------------
-                { L_,   0                  , 1                    },
-                { L_,   1                  , 2                    },
-                { L_,  -1LL                , 0                    },
-                { L_,   0xFFFFFFFFLL       , 0x100000000LL        },
-                { L_,  0xFFFFFFFFFFFFFFFFLL , 0                   }
+                //line expected                   value
+                //---- -------------------        ---------------------
+                { L_,         0                  , 1                    },
+                { L_,         1                  , 2                    },
+                { L_,        -1LL                , 0                    },
+                { L_,         0xFFFFFFFFLL       , 0x100000000LL        },
+                { L_,  (int) 0xFFFFFFFFFFFFFFFFLL , 0                   }
             };
 
             const int NUM_VALUES = sizeof VALUES / sizeof *VALUES;
@@ -3844,7 +3850,13 @@ int main(int argc, char *argv[]) {
                 { L_,   0   , 11         , 33     , 0          , 0       },
                 { L_,   1   , 19         , 1      , 19         , 1       },
                 { L_,  -1   , 4          , 1      , -1         , -1      },
-                { L_,   2   , 0xFFFFFFFF , 2      , 0xFFFFFFFF , 2       },
+                { L_,
+                  2,
+                  (int) 0xFFFFFFFF,
+                  2,
+                  (int) 0xFFFFFFFF,
+                  2
+                },
                 { L_,  -2   , 16         , 0      , -2         , -2      }
             };
 
@@ -4507,13 +4519,13 @@ int main(int argc, char *argv[]) {
                 int d_amount;      // Amount to add
                 int d_expected;    // Expected value
             } VALUES[] = {
-                //line d_base    d_amount d_expected
-                //---- --------  -------- ----------
-                { L_,   0       , -9    , -9         },
-                { L_,   1       , 0     , 1          },
-                { L_,  -1       , 1     , 0          },
-                { L_, 0xFFFFFFFF, 1     , 0          },
-                { L_,  -2       , -2    , -4         }
+                //line d_base          d_amount d_expected
+                //---- --------        -------- ----------
+                { L_,   0             , -9    , -9         },
+                { L_,   1             , 0     , 1          },
+                { L_,  -1             , 1     , 0          },
+                { L_, (int) 0xFFFFFFFF, 1     , 0          },
+                { L_,  -2             , -2    , -4         }
             };
 
             const int NUM_VALUES = sizeof VALUES / sizeof *VALUES;
